refactor(tools): decompose the DSL language-evaluation checker
check_dsl_language_evaluation.py was the last 3,000-line governance
checker (3,096 lines; single functions of ~700 and ~635 lines). Any
future edit would trip the strict Sonar gate's file and function
thresholds the way the other three checkers did.

It is now an import-stable entry point over a
tools/dsl_language_evaluation/ support package (keys, shape primitives,
claim scope/binding, measure recomputation, protocol validation with a
plans sibling, snapshot validation with a subjects/review/completion
sibling, analysis validation), every module under the 500-line cap.
_protocol_records_by_id moves to the shape module where its consumers
expect it.

Pure restructuring: the checker emits byte-identical output and exit
status on this repo in both text and --json modes, and the import
surface the test suite relies on is re-exported unchanged.

Verification: test_dsl_language_evaluation (40 tests) and the full
hermetic suite pass with the coverage gate green; ruff format and lint
clean; check_repo_policy pass.
